Refactor common utils
